[simplify] Finish(*) like term collection basics
I switched to editing the AST in-place, but this breaks other parts of
the simplifier which expect to be able to return a entirely new
expression. For example, there is no way to edit 5 + 3 in place. The
operation node needs to be wholly replaced with a new Number node.

To edit all expressions in-place, a root expression needs to be added,
like equation... so that means more retooling of the AST node family.
Ugh.